Started work on service scheduling and creating
The service model was created. It is now possible to create a service,
and to register them to a pet. The front-end for those requests was also
completed.

The route `/services/schedule` returns a list of all the scheduled
services.

Some minor improvements in the `users` route were also made.